Update style of services checkbox (#76)
This PR updates the style of the checkbox and corresponding label for the services section.
Note: this is the first PR of several to update the services section. The alignment is still a little off, but that will be fixed in future PRs that add the entire new list of services.