Load Initial Data in App (#101)
The initial data that we load should be loaded earlier when the app loads. Right now, it's loaded when you go to the verification portal. But if you only go to the mini app, it's not loaded at all. This is fixed by loading the initial data when the app component mounts. This way, we have access to it no matter which "app" we go to (mini app or verification portal). This PR makes that change and cleans up the code where it was loading before.

I also added the Google Maps API key so that we can use it in the mini app (to geocode addresses that users type in)