Disable Save button on submit (#130)
This PR disables the Save button while the form is submitting. It's important that we do that so that you can't click Submit multiple times while the form is submitting. I also made a couple of refactors. I tried to comment directly on the changed lines to explain some of the changes.