Add size parameters for tablet (#165)
We decided that it's necessary to have an additional tablet view size. This will allow us to be able to make a more responsive site. This PR just adds the additional size parameters for tablet and adds the additional code to make sure the site still works for the tablet size.